+++ src/Factory/ServerRequest/ServerRequestFactory.php
@@ -124,7 +124,7 @@ final class ServerRequestFactory implements ServerRequestFactoryInterface
             EnvironmentFactory::default(),
             CookiesFactory::default(),
             Factory\Query\QueryFactory::default(),
-            Factory\Form\FormFactory::default(),
+            FormFactory::default(),
             FilesFactory::default($io),
             $server,
         );